Move interactor to :core module
