+++ core/src/main/java/com/costular/atomtasks/core/Interactor.kt
@@ -1,9 +1,5 @@
-package com.costular.atomtasks.data
+package com.costular.atomtasks.core
 
-import com.costular.atomtasks.core.InvokeError
-import com.costular.atomtasks.core.InvokeStarted
-import com.costular.atomtasks.core.InvokeStatus
-import com.costular.atomtasks.core.InvokeSuccess
 import java.util.concurrent.TimeUnit
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.flow.Flow
